requirements.txt
numpy==1.26.4
opencv-python>=4.8.0
scipy>=1.10.0
tqdm>=4.66.0
mediapipe==0.10.14
onnxruntime>=1.16.0
protobuf>=3.20.3,<4.25.4
requests
flask
flask-socketio
paho-mqtt
eventlet
